Implemented team ownership transferring between members.

+++ BESL.Application/Teams/Commands/RemovePlayer/RemovePlayerCommandHandler.cs
@@ -37,6 +37,11 @@
                 throw new ForbiddenException();
             }
 
+            if (desiredPlayerTeamEntity.Team.OwnerId == request.PlayerId)
+            {
+                throw new OwnerOfAMustTransferOwnerShipBeforeLeavingTeamException();
+            }
+
             this.playerTeamsRepository.Delete(desiredPlayerTeamEntity);
             return await this.playerTeamsRepository.SaveChangesAsync(cancellationToken);
         }
